fix redis store tests
