+++ pkg/sessions/redis/redis_store_test.go
@@ -1,10 +1,6 @@
 package redis
 
 import (
-	"bytes"
-	"crypto/tls"
-	"encoding/pem"
-	"os"
 	"time"
 
 	"github.com/Bose/minisentinel"
@@ -13,42 +9,14 @@ import (
 	sessionsapi "github.com/oauth2-proxy/oauth2-proxy/v7/pkg/apis/sessions"
 	"github.com/oauth2-proxy/oauth2-proxy/v7/pkg/sessions/persistence"
 	"github.com/oauth2-proxy/oauth2-proxy/v7/pkg/sessions/tests"
-	"github.com/oauth2-proxy/oauth2-proxy/v7/pkg/util"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 )
 
-const redisUsername = "testuser"
-const redisPassword = "0123456789abcdefghijklmnopqrstuv"
-
-var (
-	cert   tls.Certificate
-	caPath string
+const (
+	redisUsername = "testuser"
+	redisPassword = "0123456789abcdefghijklmnopqrstuv"
 )
-
-var _ = BeforeSuite(func() {
-	var err error
-	certBytes, keyBytes, err := util.GenerateCert("127.0.0.1")
-	Expect(err).ToNot(HaveOccurred())
-	certOut := new(bytes.Buffer)
-	Expect(pem.Encode(certOut, &pem.Block{Type: "CERTIFICATE", Bytes: certBytes})).To(Succeed())
-	certData := certOut.Bytes()
-	keyOut := new(bytes.Buffer)
-	Expect(pem.Encode(keyOut, &pem.Block{Type: "PRIVATE KEY", Bytes: keyBytes})).To(Succeed())
-	cert, err = tls.X509KeyPair(certData, keyOut.Bytes())
-	Expect(err).ToNot(HaveOccurred())
-
-	certFile, err := os.CreateTemp("", "cert.*.pem")
-	Expect(err).ToNot(HaveOccurred())
-	caPath = certFile.Name()
-	_, err = certFile.Write(certData)
-	defer certFile.Close()
-	Expect(err).ToNot(HaveOccurred())
-})
-
-var _ = AfterSuite(func() {
-	Expect(os.Remove(caPath)).ToNot(HaveOccurred())
-})
 
 var _ = Describe("Redis SessionStore Tests", func() {
 	// helper interface to allow us to close client connections
@@ -103,10 +71,6 @@ var _ = Describe("Redis SessionStore Tests", func() {
 		BeforeEach(func() {
 			ms = minisentinel.NewSentinel(mr)
 			Expect(ms.Start()).To(Succeed())
-		})
-
-		AfterEach(func() {
-			ms.Close()
 		})
 
 		tests.RunSessionStoreTests(
@@ -183,10 +147,6 @@ var _ = Describe("Redis SessionStore Tests", func() {
 			BeforeEach(func() {
 				ms = minisentinel.NewSentinel(mr)
 				Expect(ms.Start()).To(Succeed())
-			})
-
-			AfterEach(func() {
-				ms.Close()
 			})
 
 			tests.RunSessionStoreTests(
@@ -281,192 +241,5 @@ var _ = Describe("Redis SessionStore Tests", func() {
 				},
 			)
 		})
-	})
-
-	const redissProtocol = "rediss://"
-	Context("with TLS connection", func() {
-		BeforeEach(func() {
-			mr.Close()
-
-			var err error
-			mr, err = miniredis.RunTLS(&tls.Config{Certificates: []tls.Certificate{cert}})
-			Expect(err).ToNot(HaveOccurred())
-		})
-		AfterEach(func() {
-			mr.Close()
-
-			var err error
-			mr, err = miniredis.Run()
-			Expect(err).ToNot(HaveOccurred())
-		})
-
-		Context("with standalone", func() {
-			tests.RunSessionStoreTests(
-				func(opts *options.SessionOptions, cookieOpts *options.Cookie) (sessionsapi.SessionStore, error) {
-					// Set the connection URL
-					opts.Type = options.RedisSessionStoreType
-					opts.Redis.ConnectionURL = redissProtocol + mr.Addr()
-					opts.Redis.CAPath = caPath
-
-					// Capture the session store so that we can close the client
-					ss, err := NewRedisSessionStore(opts, cookieOpts)
-					return ss, err
-				},
-				func(d time.Duration) error {
-					mr.FastForward(d)
-					return nil
-				},
-			)
-		})
-
-		Context("with cluster", func() {
-			tests.RunSessionStoreTests(
-				func(opts *options.SessionOptions, cookieOpts *options.Cookie) (sessionsapi.SessionStore, error) {
-					clusterAddr := redissProtocol + mr.Addr()
-					opts.Type = options.RedisSessionStoreType
-					opts.Redis.ClusterConnectionURLs = []string{clusterAddr}
-					opts.Redis.UseCluster = true
-					opts.Redis.CAPath = caPath
-
-					// Capture the session store so that we can close the client
-					var err error
-					ss, err = NewRedisSessionStore(opts, cookieOpts)
-					return ss, err
-				},
-				func(d time.Duration) error {
-					mr.FastForward(d)
-					return nil
-				},
-			)
-		})
-	})
-
-	Context("with insecure TLS connection", func() {
-		BeforeEach(func() {
-			mr.Close()
-
-			var err error
-			mr, err = miniredis.RunTLS(&tls.Config{Certificates: []tls.Certificate{cert}})
-			Expect(err).ToNot(HaveOccurred())
-		})
-		AfterEach(func() {
-			mr.Close()
-
-			var err error
-			mr, err = miniredis.Run()
-			Expect(err).ToNot(HaveOccurred())
-		})
-
-		Context("with standalone", func() {
-			tests.RunSessionStoreTests(
-				func(opts *options.SessionOptions, cookieOpts *options.Cookie) (sessionsapi.SessionStore, error) {
-					// Set the connection URL
-					opts.Type = options.RedisSessionStoreType
-					opts.Redis.ConnectionURL = redissProtocol + mr.Addr()
-					opts.Redis.InsecureSkipTLSVerify = true
-
-					// Capture the session store so that we can close the client
-					ss, err := NewRedisSessionStore(opts, cookieOpts)
-					return ss, err
-				},
-				func(d time.Duration) error {
-					mr.FastForward(d)
-					return nil
-				},
-			)
-		})
-
-		Context("with cluster", func() {
-			tests.RunSessionStoreTests(
-				func(opts *options.SessionOptions, cookieOpts *options.Cookie) (sessionsapi.SessionStore, error) {
-					clusterAddr := redissProtocol + mr.Addr()
-					opts.Type = options.RedisSessionStoreType
-					opts.Redis.ClusterConnectionURLs = []string{clusterAddr}
-					opts.Redis.UseCluster = true
-					opts.Redis.InsecureSkipTLSVerify = true
-
-					// Capture the session store so that we can close the client
-					var err error
-					ss, err = NewRedisSessionStore(opts, cookieOpts)
-					return ss, err
-				},
-				func(d time.Duration) error {
-					mr.FastForward(d)
-					return nil
-				},
-			)
-		})
-	})
-
-	Context("with custom CA path", func() {
-		BeforeEach(func() {
-			mr.Close()
-
-			var err error
-			mr, err = miniredis.RunTLS(&tls.Config{Certificates: []tls.Certificate{cert}})
-			Expect(err).ToNot(HaveOccurred())
-		})
-
-		AfterEach(func() {
-			mr.Close()
-
-			var err error
-			mr, err = miniredis.Run()
-			Expect(err).ToNot(HaveOccurred())
-		})
-
-		tests.RunSessionStoreTests(
-			func(opts *options.SessionOptions, cookieOpts *options.Cookie) (sessionsapi.SessionStore, error) {
-				// Set the connection URL
-				opts.Type = options.RedisSessionStoreType
-				opts.Redis.ConnectionURL = redisProtocol + mr.Addr()
-				opts.Redis.CAPath = caPath
-
-				// Capture the session store so that we can close the client
-				var err error
-				ss, err = NewRedisSessionStore(opts, cookieOpts)
-				return ss, err
-			},
-			func(d time.Duration) error {
-				mr.FastForward(d)
-				return nil
-			},
-		)
-	})
-
-	Context("with insecure TLS connection", func() {
-		BeforeEach(func() {
-			mr.Close()
-
-			var err error
-			mr, err = miniredis.RunTLS(&tls.Config{Certificates: []tls.Certificate{cert}})
-			Expect(err).ToNot(HaveOccurred())
-		})
-
-		AfterEach(func() {
-			mr.Close()
-
-			var err error
-			mr, err = miniredis.Run()
-			Expect(err).ToNot(HaveOccurred())
-		})
-
-		tests.RunSessionStoreTests(
-			func(opts *options.SessionOptions, cookieOpts *options.Cookie) (sessionsapi.SessionStore, error) {
-				// Set the connection URL
-				opts.Type = options.RedisSessionStoreType
-				opts.Redis.ConnectionURL = "redis://127.0.0.1:" + mr.Port() // func (*Miniredis) StartTLS listens on 127.0.0.1
-				opts.Redis.InsecureSkipTLSVerify = true
-
-				// Capture the session store so that we can close the client
-				var err error
-				ss, err = NewRedisSessionStore(opts, cookieOpts)
-				return ss, err
-			},
-			func(d time.Duration) error {
-				mr.FastForward(d)
-				return nil
-			},
-		)
 	})
 })
